move status stuff to types and utils
