+++ client/src/components/v3/status/WMStatusContainer.tsx
@@ -9,8 +9,6 @@ import WMStatus, { WMStatusProps } from 'components/v3/status/WMStatus';
  * @returns Component
  */
 export default function WMStatusContainer() {
-  // TODO: Link API
-
   const front: WMStatusProps = {
     moduleName: 'Front WM',
     online: false,
